scripts/requirements.txt
kaggle==1.6.17
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0
pandas==2.2.3
pyarrow==17.0.0
python-dotenv==1.0.1
